chore: addressed duplication issues
